Clean up example project
